requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3

# Image Processing
numpy==1.26.1
opencv-python==4.8.1.78
Pillow==10.1.0

# OCR
easyocr==1.7.1

# Frontend
streamlit==1.28.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0